Use round, roundf, roundl from C++ std library. These functions are guaranteed to be part of C++11 standard library.
* configure.ac: Remove AC_CHECK_FUNCS, AC_CHECK_DECLS for round.

* lo-mappers.cc: Delegate to std::round or std::roundf.

* oct-inttypes.h: #include <cmath>.  Delegate to std::roundl.

+++ liboctave/numeric/lo-mappers.cc
@@ -293,7 +293,7 @@ namespace octave
     float floor (float x) { return std::floor (x); }
 
     double round (double x) { return std::round (x); }
-    float round (float x) { return std::round (x); }
+    float round (float x) { return std::roundf (x); }
 
     double frexp (double x, int *expptr)
     {
